combining TagInput and RecordTagInput into one class
