+++ src/app/components/RecordTagInput/base.less
@@ -2,4 +2,14 @@
   .ui.search .prompt {
     border-radius: 4px;
   }
+  .tag-container {
+    .ui.label {
+      margin-bottom: 5px;
+    }
+  }
+  .input-container {
+    .ui.input {
+      margin-top: 5px;
+    }
+  }
 }
