+++ src/app/components/RecordTagInput/base.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import {Search, Icon} from 'semantic-ui-react';
-import {TagInput} from 'components/TagInput';
+import {Search, Icon, Label} from 'semantic-ui-react';
 import './base.less';
 import {ITagResult, getTags} from 'apollo/modules/tags';
+import * as TagsInput from 'react-tagsinput';
 
 interface IProps {
   onChange: (tags: string[]) => void;
@@ -26,6 +26,38 @@ class RecordTagInputInner extends React.Component<IProps, any> {
   constructor(props) {
     super(props);
     this.renderInput = this.renderInput.bind(this);
+    this.renderTag = this.renderTag.bind(this);
+  }
+
+  private renderTag(props) {
+    const {tag, key, disabled, onRemove, getTagDisplayValue} = props;
+    const onRemoveLocal = () => {
+      if (disabled) {
+        return;
+      }
+      onRemove(key);
+    };
+    return (
+      <Label key={key} as="a" onClick={onRemoveLocal}>
+        <span>{getTagDisplayValue(tag)}</span>
+        {!disabled &&
+        <Icon name="close"/>
+        }
+      </Label>
+    );
+  }
+
+  private renderLayout(tagComponents, inputComponent) {
+    return (
+      <div className="tag-input-container">
+        <div className="tag-container">
+          {tagComponents}
+        </div>
+        <div className="input-container">
+          {inputComponent}
+        </div>
+      </div>
+    );
   }
 
   private renderInput(props) {
@@ -65,17 +97,24 @@ class RecordTagInputInner extends React.Component<IProps, any> {
   }
 
   public render() {
+    const addOnBlur = false;
+    const addKeys = [];
     return (
       <div className="record-tag-input">
-        <TagInput
+        <TagsInput
           id={this.props.id}
+          className="tag-input"
+          value={this.props.tags}
           onChange={this.props.onChange}
+          renderTag={this.renderTag}
           renderInput={this.renderInput}
-          addOnBlur={false}
-          addOnKeyboard={false}
-          tags={this.props.tags}
+          renderLayout={this.renderLayout}
+          addOnBlur={addOnBlur}
+          removeKeys={[]}
+          addKeys={addKeys}
         />
       </div>
+
     );
   }
 }
