--- src/app/components/TagInput/index.tsx
@@ -1,93 +1,0 @@
-import * as React from 'react';
-import {Label, Icon, Input} from 'semantic-ui-react';
-import * as TagsInput from 'react-tagsinput';
-import './style.less';
-
-interface IProps {
-  onChange: (tags: string[]) => void;
-  renderInput?: (props) => JSX.Element;
-  addOnBlur?: boolean;
-  addOnKeyboard?: boolean;
-  tags: string[];
-  id?: string;
-}
-
-class TagInput extends React.Component<IProps, any> {
-
-  constructor(props) {
-    super(props);
-    this.onChange = this.onChange.bind(this);
-    this.renderTag = this.renderTag.bind(this);
-    this.renderInput = this.renderInput.bind(this);
-  }
-
-  private onChange(tags: string[]): void {
-    this.props.onChange(tags);
-  }
-
-  private renderTag(props) {
-    const {tag, key, disabled, onRemove, getTagDisplayValue} = props;
-    const onRemoveLocal = () => {
-      if (disabled) {
-        return;
-      }
-      onRemove(key);
-    };
-    return (
-      <Label key={key} as="a" onClick={onRemoveLocal}>
-        <span>{getTagDisplayValue(tag)}</span>
-        {!disabled &&
-          <Icon name="close"/>
-        }
-      </Label>
-    );
-  }
-
-  private renderInput(props) {
-    const {onChange, value, ...other} = props;
-    return (
-      <Input icon={true} type="text" placeholder="Add tag" onChange={onChange} value={value} {...other}>
-        <input />
-        <Icon name="add" link={true} />
-      </Input>
-    );
-  }
-
-  private renderLayout(tagComponents, inputComponent) {
-    return (
-      <div className="tag-input-container">
-        <div className="tag-container">
-          {tagComponents}
-        </div>
-        <div className="input-container">
-          {inputComponent}
-        </div>
-      </div>
-    );
-  }
-
-  public render() {
-    const input = this.props.renderInput || this.renderInput;
-    const addOnBlur = this.props.addOnBlur !== false;
-    const addKeys = this.props.addOnKeyboard === false ? [] : [13]; // enter
-    return (
-      <div>
-        <TagsInput
-          id={this.props.id}
-          className="tag-input"
-          value={this.props.tags}
-          onChange={this.onChange}
-          renderTag={this.renderTag}
-          renderInput={input}
-          renderLayout={this.renderLayout}
-          addOnBlur={addOnBlur}
-          removeKeys={[]}
-          addKeys={addKeys}
-        />
-      </div>
-
-    );
-  }
-}
-
-export { TagInput };
