--- src/app/components/TagInput/style.less
@@ -1,13 +1,0 @@
-.tag-input {
-  .tag-container {
-    .ui.label {
-      margin-bottom: 5px;
-    }
-  }
-  .input-container {
-    .ui.input {
-      margin-top: 5px;
-    }
-  }
-}
-
